break out HistoryRow, add drawer
